Removed the order property from metric entry
It didn't make a ton of sense and would have required the entity to be
dependent on another without any hard ties. Would be better to have a
created time but that adds extra memory that's not asked for by doc.

Also moved the sync block from the controller to the store where it
could be handled better by a better store.